Smart Energy Challenge/requirements.txt
pyserial>=3.1.1
pygsheets>=2.0.5
pandas>=1.5.0